# backend/api/gen/go/admin/service/v1/i_operation_audit_log_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             (unknown)
// source: admin/service/v1/i_operation_audit_log.proto

package adminpb

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
	v1 "github.com/tx7do/go-crud/api/gen/go/pagination/v1"
	v11 "go-wind-admin/api/gen/go/audit/service/v1"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationOperationAuditLogServiceGet = "/admin.service.v1.OperationAuditLogService/Get"
const OperationOperationAuditLogServiceList = "/admin.service.v1.OperationAuditLogService/List"

type OperationAuditLogServiceHTTPServer interface {
	// Get 查询操作审计日志详情
	Get(context.Context, *v11.GetOperationAuditLogRequest) (*v11.OperationAuditLog, error)
	// List 查询操作审计日志列表
	List(context.Context, *v1.PagingRequest) (*v11.ListOperationAuditLogResponse, error)
}

func RegisterOperationAuditLogServiceHTTPServer(s *http.Server, srv OperationAuditLogServiceHTTPServer) {
	r := s.Route("/")
	r.GET("/admin/v1/operation-audit-logs", _OperationAuditLogService_List9_HTTP_Handler(srv))
	r.GET("/admin/v1/operation-audit-logs/{id}", _OperationAuditLogService_Get9_HTTP_Handler(srv))
}

func _OperationAuditLogService_List9_HTTP_Handler(srv OperationAuditLogServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v1.PagingRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationOperationAuditLogServiceList)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.List(ctx, req.(*v1.PagingRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.ListOperationAuditLogResponse)
		return ctx.Result(200, reply)
	}
}

func _OperationAuditLogService_Get9_HTTP_Handler(srv OperationAuditLogServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in v11.GetOperationAuditLogRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationOperationAuditLogServiceGet)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Get(ctx, req.(*v11.GetOperationAuditLogRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*v11.OperationAuditLog)
		return ctx.Result(200, reply)
	}
}

type OperationAuditLogServiceHTTPClient interface {
	// Get 查询操作审计日志详情
	Get(ctx context.Context, req *v11.GetOperationAuditLogRequest, opts ...http.CallOption) (rsp *v11.OperationAuditLog, err error)
	// List 查询操作审计日志列表
	List(ctx context.Context, req *v1.PagingRequest, opts ...http.CallOption) (rsp *v11.ListOperationAuditLogResponse, err error)
}

type OperationAuditLogServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewOperationAuditLogServiceHTTPClient(client *http.Client) OperationAuditLogServiceHTTPClient {
	return &OperationAuditLogServiceHTTPClientImpl{client}
}

// Get 查询操作审计日志详情
func (c *OperationAuditLogServiceHTTPClientImpl) Get(ctx context.Context, in *v11.GetOperationAuditLogRequest, opts ...http.CallOption) (*v11.OperationAuditLog, error) {
	var out v11.OperationAuditLog
	pattern := "/admin/v1/operation-audit-logs/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationOperationAuditLogServiceGet))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// List 查询操作审计日志列表
func (c *OperationAuditLogServiceHTTPClientImpl) List(ctx context.Context, in *v1.PagingRequest, opts ...http.CallOption) (*v11.ListOperationAuditLogResponse, error) {
	var out v11.ListOperationAuditLogResponse
	pattern := "/admin/v1/operation-audit-logs"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationOperationAuditLogServiceList))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}