# backend/app/admin/service/cmd/server/assets/openapi.yaml
...
openapi: 3.0.3
info:
    title: GoWind Admin API
    description: GoWind Admin API
    contact:
        name: tx7do
        url: https://github.com/tx7do/go-wind-admin
        email: yanglinbo@gmail.com
    license:
        name: MIT License
        url: https://github.com/tx7do/go-wind-admin/blob/master/LICENSE
    version: "1.0"
paths:
...
    /admin/v1/login:
        post:
            tags:
                - AuthenticationService
            description: 登录
            operationId: AuthenticationService_Login
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/LoginRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/LoginResponse'
            security:
                - {}
...
    /admin/v1/login-policies:
        get:
            tags:
                - LoginPolicyService
            description: 查询登录策略列表
            operationId: LoginPolicyService_List
            parameters:
                - name: page
                  in: query
                  description: 当前页码（从1开始，默认1）
                  schema:
                    type: integer
                    format: uint32
                - name: pageSize
                  in: query
                  description: 每页条数（默认10，建议设置上限如100）
                  schema:
                    type: integer
                    format: uint32
                - name: offset
                  in: query
                  description: 跳过的记录数（从0开始，默认0）
                  schema:
                    type: string
                - name: limit
                  in: query
                  description: 最多返回的记录数（默认10，建议设置上限如100）
                  schema:
                    type: integer
                    format: uint32
                - name: token
                  in: query
                  description: 上一页最后一条记录的游标（如ID/时间戳+ID，首次请求为空）
                  schema:
                    type: string
                - name: noPaging
                  in: query
                  description: 是否不分页，如果为true，则page和pageSize参数无效。
                  schema:
                    type: boolean
                - name: query
                  in: query
                  description: JSON字符串过滤条件，基础语法：{"field1":"val1", "field2___icontains":"val2"}，具体请参见：https://github.com/tx7do/go-crud/tree/main/pagination/filter/README.md
                  schema:
                    type: string
                - name: filter
                  in: query
                  description: Google AIP规范字符串过滤条件
                  schema:
                    type: string
                - name: filterExpr.type
                  in: query
                  description: 过滤表达式类型
                  schema:
                    enum:
                        - EXPR_TYPE_UNSPECIFIED
                        - AND
                        - OR
                    type: string
                    format: enum
                - name: orderBy
                  in: query
                  description: 排序条件
                  schema:
                    type: string
                - name: fieldMask
                  in: query
                  description: 字段掩码，其作用为SELECT中的字段，其语法为使用逗号分隔字段名，例如：id,realName,userName。如果为空则选中所有字段，即SELECT *。
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListLoginPolicyResponse'
        post:
            tags:
                - LoginPolicyService
            description: 创建登录策略
            operationId: LoginPolicyService_Create
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateLoginPolicyRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /admin/v1/login-policies/{id}:
        get:
            tags:
                - LoginPolicyService
            description: 查询登录策略详情
            operationId: LoginPolicyService_Get
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
                - name: viewMask
                  in: query
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/LoginPolicy'
        put:
            tags:
                - LoginPolicyService
            description: 更新登录策略
            operationId: LoginPolicyService_Update
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateLoginPolicyRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
        delete:
            tags:
                - LoginPolicyService
            description: 删除登录策略
            operationId: LoginPolicyService_Delete
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content: {}
...
    /admin/v1/operation-audit-logs:
        get:
            tags:
                - OperationAuditLogService
            description: 查询操作审计日志列表
            operationId: OperationAuditLogService_List
            parameters:
                - name: page
                  in: query
                  description: 当前页码（从1开始，默认1）
                  schema:
                    type: integer
                    format: uint32
                - name: pageSize
                  in: query
                  description: 每页条数（默认10，建议设置上限如100）
                  schema:
                    type: integer
                    format: uint32
                - name: offset
                  in: query
                  description: 跳过的记录数（从0开始，默认0）
                  schema:
                    type: string
                - name: limit
                  in: query
                  description: 最多返回的记录数（默认10，建议设置上限如100）
                  schema:
                    type: integer
                    format: uint32
                - name: token
                  in: query
                  description: 上一页最后一条记录的游标（如ID/时间戳+ID，首次请求为空）
                  schema:
                    type: string
                - name: noPaging
                  in: query
                  description: 是否不分页，如果为true，则page和pageSize参数无效。
                  schema:
                    type: boolean
                - name: query
                  in: query
                  description: JSON字符串过滤条件，基础语法：{"field1":"val1", "field2___icontains":"val2"}，具体请参见：https://github.com/tx7do/go-crud/tree/main/pagination/filter/README.md
                  schema:
                    type: string
                - name: filter
                  in: query
                  description: Google AIP规范字符串过滤条件
                  schema:
                    type: string
                - name: filterExpr.type
                  in: query
                  description: 过滤表达式类型
                  schema:
                    enum:
                        - EXPR_TYPE_UNSPECIFIED
                        - AND
                        - OR
                    type: string
                    format: enum
                - name: orderBy
                  in: query
                  description: 排序条件
                  schema:
                    type: string
                - name: fieldMask
                  in: query
                  description: 字段掩码，其作用为SELECT中的字段，其语法为使用逗号分隔字段名，例如：id,realName,userName。如果为空则选中所有字段，即SELECT *。
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListOperationAuditLogResponse'
    /admin/v1/operation-audit-logs/{id}:
        get:
            tags:
                - OperationAuditLogService
            description: 查询操作审计日志详情
            operationId: OperationAuditLogService_Get
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
                - name: viewMask
                  in: query
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/OperationAuditLog'
...
components:
    schemas:
...
        CreateLoginPolicyRequest:
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/LoginPolicy'
            description: 创建登录策略 - 请求
...
        GeoLocation:
            type: object
            properties:
                countryCode:
                    type: string
                    description: 国家代码（CN/US），遵循ISO 3166-1 alpha-2标准
                province:
                    type: string
                    description: 省份/州（中文：北京市，英文：California）
                city:
                    type: string
                    description: 城市（中文：上海市，英文：New York）
                isp:
                    type: string
                    description: 网络运营商（如：中国移动/中国电信/AT&T）
                addressRemark:
                    type: string
                    description: 地址备注（如：内网IP/无法解析/海外IP）
                latitude:
                    type: string
                    description: 纬度（微度）
                longitude:
                    type: string
                    description: 经度（微度）
            description: 地理位置
...
        ListLoginPolicyResponse:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/LoginPolicy'
                total:
                    type: string
            description: 查询登录策略列表 - 回应
...
        ListOperationAuditLogResponse:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/OperationAuditLog'
                total:
                    type: string
            description: 查询操作审计日志列表 - 回应
...
        LoginPolicy:
            type: object
            properties:
                id:
                    type: integer
                    description: 登录策略ID
                    format: uint32
                targetId:
                    type: integer
                    description: 目标用户ID
                    format: uint32
                type:
                    enum:
                        - LOGIN_RESTRICTION_TYPE_UNSPECIFIED
                        - BLACKLIST
                        - WHITELIST
                    type: string
                    description: 限制类型
                    format: enum
                method:
                    enum:
                        - LOGIN_RESTRICTION_METHOD_UNSPECIFIED
                        - IP
                        - MAC
                        - REGION
                        - TIME
                        - DEVICE
                    type: string
                    description: 限制方式
                    format: enum
                value:
                    type: string
                    description: 限制值（如IP地址、MAC地址或地区代码）
                reason:
                    type: string
                    description: 限制原因
                tenantId:
                    type: integer
                    description: 租户ID，0代表系统全局角色
                    format: uint32
                tenantName:
                    type: string
                    description: 租户名称
                createdBy:
                    type: integer
                    description: 创建者ID
                    format: uint32
                updatedBy:
                    type: integer
                    description: 更新者ID
                    format: uint32
                deletedBy:
                    type: integer
                    description: 删除者用户ID
                    format: uint32
                createdAt:
                    type: string
                    description: 创建时间
                    format: date-time
                updatedAt:
                    type: string
                    description: 更新时间
                    format: date-time
                deletedAt:
                    type: string
                    description: 删除时间
                    format: date-time
            description: 登录策略
        LoginRequest:
            required:
                - grant_type
            type: object
            properties:
                grant_type:
                    enum:
                        - password
                        - client_credentials
                        - authorization_code
                        - refresh_token
                        - implicit
                    type: string
                    default: password
                    description: 授权类型，此处的值固定为"password"，必选项。
                    format: enum
                client_id:
                    type: string
                    description: 客户端ID
                client_secret:
                    type: string
                    description: 客户端密钥
                scope:
                    type: string
                    description: 以空格分隔的用户授予范围列表。如果未提供，scope则授权任何范围，默认为空列表。
                redirect_uri:
                    type: string
                    description: 跳转链接
                username:
                    type: string
                    description: 用户名
                password:
                    type: string
                    description: 用户的密码
                user_id:
                    type: integer
                    description: 用户ID
                    format: uint32
                refresh_token:
                    type: string
                    description: 更新令牌，用来获取下一次的访问令牌，可选项。如果访问令牌将过期，则返回刷新令牌很有用，应用程序可以使用该刷新令牌来获取另一个访问令牌。但是，通过隐式授予颁发的令牌不能颁发刷新令牌。
                code:
                    type: string
                    description: 授权请求中收到的一次性验证/认证码。(当使用授权码模式时)
                client_type:
                    enum:
                        - admin
                        - app
                    type: string
                    description: 客户端类型
                    format: enum
                device_id:
                    type: string
                    description: 设备唯一标识（可选），用于设备绑定、推送、风控等
            description: 用户后台登录 - 请求
        LoginResponse:
            type: object
            properties:
                access_token:
                    type: string
                    description: 访问令牌，必选项。授权服务器颁发的访问令牌字符串。
                refresh_token:
                    type: string
                    description: 更新令牌，用来获取下一次的访问令牌，可选项。如果访问令牌将过期，则返回刷新令牌很有用，应用程序可以使用该刷新令牌来获取另一个访问令牌。但是，通过隐式授予颁发的令牌不能颁发刷新令牌。
                token_type:
                    enum:
                        - bearer
                        - mac
                    type: string
                    default: Bearer
                    description: 令牌的类型，该值大小写不敏感，必选项，可以是bearer类型或mac类型，通常只是字符串“Bearer”。
                    format: enum
                expires_in:
                    type: string
                    description: 令牌有效时间，单位为秒。如果访问令牌过期，服务器应回复授予访问令牌的持续时间。如果省略该参数，必须其他方式设置过期时间。
                scope:
                    type: string
                    description: 以空格分隔的用户授予范围列表。如果未提供，scope则授权任何范围，默认为空列表。
            description: 用户后台登录 - 回应
...
        OperationAuditLog:
            type: object
            properties:
                id:
                    type: integer
                    description: API审计日志ID
                    format: uint32
                tenantId:
                    type: integer
                    description: 租户ID
                    format: uint32
                tenantName:
                    type: string
                    description: 租户名称
                userId:
                    type: integer
                    description: 用户ID
                    format: uint32
                username:
                    type: string
                    description: 账号名
                resourceType:
                    type: string
                    description: 资源类型
                resourceId:
                    type: string
                    description: 资源ID
                action:
                    enum:
                        - ACTION_TYPE_UNSPECIFIED
                        - CREATE
                        - UPDATE
                        - DELETE
                        - READ
                        - ASSIGN
                        - UNASSIGN
                        - EXPORT
                        - OTHER
                    type: string
                    description: 动作
                    format: enum
                beforeData:
                    type: string
                    description: 操作前数据
                afterData:
                    type: string
                    description: 操作后数据
                sensitiveLevel:
                    enum:
                        - SENSITIVE_LEVEL_UNSPECIFIED
                        - PUBLIC
                        - INTERNAL
                        - CONFIDENTIAL
                        - SECRET
                    type: string
                    description: 敏感等级
                    format: enum
                requestId:
                    type: string
                    description: 全局请求ID（关联网关日志）
                traceId:
                    type: string
                    description: 全局链路追踪ID（符合W3C TraceContext标准）
                success:
                    type: boolean
                    description: 操作是否成功
                failureReason:
                    type: string
                    description: 失败原因
                ipAddress:
                    type: string
                    description: IP地址
                geoLocation:
                    $ref: '#/components/schemas/GeoLocation'
                logHash:
                    type: string
                    description: 日志内容哈希（SHA256，十六进制字符串）
                signature:
                    type: string
                    description: 日志数字签名（ECDSA，签名内容：tenant_id+user_id+created_at+log_hash）
                    format: bytes
                createdAt:
                    type: string
                    description: 日志创建时间
                    format: date-time
            description: 操作审计日志
...
        UpdateLoginPolicyRequest:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                data:
                    $ref: '#/components/schemas/LoginPolicy'
                updateMask:
                    example: id,realname,username
                    type: string
                    description: 要更新的字段列表
                    format: field-mask
                allowMissing:
                    type: boolean
                    description: 如果设置为true的时候，资源不存在则会新增(插入)，并且在这种情况下`updateMask`字段将会被忽略。
            description: 更新登录策略 - 请求
...
    securitySchemes:
        OAuth2PasswordBearer:
            type: oauth2
            flows:
                password:
                    tokenUrl: /admin/v1/login
                    refreshUrl: /admin/v1/refresh-token
                    scopes: {}
security:
    - OAuth2PasswordBearer: []
...